+++ private/plugins/forum/language/english_utf-8.php
@@ -458,6 +458,8 @@ $LANG_GF02 = array(
     'invalid_email' => 'Email empty or not a valid email address',
     'invalid_name'  => 'No name or name blank',
     'last_edited'   => 'Last Edited',
+    'copy_to_clipboard' => 'Copy Post Permalink to Clipboard',
+    'post_ctc'      => 'Permalink Copied to Clipboard',
 );
 
 $LANG_GF03 = array(