+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v1.6.7 (unreleased)
+  - Forum - Add copy to clipboard for post permalink
+  - Forum - Added tooltip (hover) for bookmark / remove bookmark
   - Media Gallery - new copy to clipboard for media id
   - Update ckeditor to v4.7.0
   - CKeditor - auto enable (start) SCAYT (Spell Check As You Type)